Reorganized folder structure and added icons to existing levels
