+++ Mockups/BuilderGUI/resources/Hexominoes_Images/3.png
