+++ Mockups/BuilderGUI/resources/Hexominoes_Images/6.png
