+++ Mockups/BuilderGUI/resources/Icons/Redo.png
