+++ Mockups/BuilderGUI/resources/Icons/RotateLeft.png
