+++ Mockups/BuilderGUI/resources/Icons/RotateRight.png
